Reworked the validation data selection workflow.
Since the input data for the SRVs are timeseries, randomly splitting the data for training and validation is not an option. The dataset creation workflow now supports three common cases for automatic train/validation data spilitting as well as the allowing the user to feed the validation and train data manually.
If  is set to zero during model definition, all data fed to the  function is treated as training data and no validation will be performed. The user can choose to manually feed the validation data to the  function using the  keyword.
If  is non-zero, one of the three cases can happen:
1) If the input data is a single  (a single continuous trajectory), the last  of the data is separated and used for validation.
2) If the input data is a  of s:
    a) If all the trajectories are the same length,  of the trajectories are used for validation. E.g, if a list of 5 trajectories of the same length are given as input and  is set to , the last trajectory in the list will be used for validation. If  is not an integer, it will be rounded up. E.g, list of 5 trajectories with  set to  will use 2 of the 5 trajectories (without splitting any of the trajectories) for validation (this means  will be automatically adjusted to  by the code).
    b)If all the trajectories are not the same length, the shortest trajectory is used for validation.
The above default behavior can be overridden by the used through  and the  keyword.